test: cover catalog setup gate
