+++ sonar-project.properties
@@ -17,7 +17,7 @@ sonar.kotlin.detekt.reportPaths=backend/build/reports/detekt/detekt.xml
 sonar.typescript.tsconfigPaths=dashboard/tsconfig.json
 
 # Framework wiring and blocking ingestion loops are integration-heavy; core metric behavior is covered separately.
-sonar.coverage.exclusions=**/com/moneat/**/models/**,**/com/moneat/config/**,**/com/moneat/plugins/**,**/com/moneat/logging/**,**/com/moneat/enterprise/**,**/com/moneat/sso/**,**/com/moneat/Application*.kt,**/com/moneat/di/**,**/com/moneat/monitoring/MonitoringModule*.kt,**/*IngestionWorker*.kt,**/com/moneat/workflows/services/WorkflowExecutionWorker.kt,**/com/moneat/billing/services/BillingBackgroundService.kt,**/com/moneat/utils/WorkerQueueSupport.kt,dashboard/src/routeTree.gen.ts,dashboard/src/components/monitoring/catalog/resourceCatalogData.ts
+sonar.coverage.exclusions=**/com/moneat/**/models/**,**/com/moneat/config/**,**/com/moneat/plugins/**,**/com/moneat/logging/**,**/com/moneat/enterprise/**,**/com/moneat/sso/**,**/com/moneat/Application*.kt,**/com/moneat/di/**,**/com/moneat/monitoring/MonitoringModule*.kt,**/*IngestionWorker*.kt,**/com/moneat/workflows/services/WorkflowExecutionWorker.kt,**/com/moneat/billing/services/BillingBackgroundService.kt,**/com/moneat/utils/WorkerQueueSupport.kt,backend/src/main/kotlin/com/moneat/monitor/services/ResourceCatalogQueryClient.kt,dashboard/src/routeTree.gen.ts,dashboard/src/components/monitoring/catalog/resourceCatalogData.ts
 sonar.cpd.exclusions=dashboard/src/components/monitoring/catalog/resourceCatalogData.ts
 
 # Ignore kotlin:S100 (function naming) in test files - Kotlin uses backtick test names
